requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.7
click==8.0.3
colorama==0.4.4
decorator==4.4.2
Flask==2.0.2
idna==3.3
imageio==2.11.1
imageio-ffmpeg==0.4.5
itsdangerous==2.0.1
Jinja2==3.0.3
MarkupSafe==2.0.1
moviepy==1.0.3
numpy==1.21.4
Pillow==8.4.0
proglog==0.1.9
python-dotenv==0.19.2
requests==2.26.0
tqdm==4.62.3
urllib3==1.26.7
Werkzeug==2.0.2
